requirements.txt
absl-py==2.1.0
aiofiles==22.1.0
aiosqlite==0.19.0
alembic==1.12.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
cached-property==1.5.2
cffi==1.15.1
charset-normalizer==3.3.2
colorlog==6.8.2
comm==0.1.4
conda-pack==0.7.1
cramjam==2.8.1
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.0
fastjsonschema==2.19.1
fastparquet==0.8.1
fonttools==4.38.0
fqdn==1.5.1
fsspec==2023.1.0
greenlet==3.0.3
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.16
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_server_fileid==0.9.1
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lightgbm==3.2.1
Mako==1.2.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
notebook==6.5.6
notebook_shim==0.2.4
numpy==1.21.6
optuna==3.5.0
ortools==9.5.2237
packaging==23.2
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.3
pathlib==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.43
protobuf==4.24.4
psutil==5.9.8
ptyprocess==0.7.0
pyarrow==12.0.1
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
pyrsistent==0.19.3
pytest==7.4.4
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==0.24.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.27
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.66.2
traitlets==5.9.0
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.10
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.15.0
